feat: update vutils, axis support to use lastScale
