+++ packages/vrender-components/src/axis/base.ts
@@ -40,9 +40,12 @@ import { getElMap, getVerticalCoord } from './util';
 import { dispatchClickState, dispatchHoverState, dispatchUnHoverState } from '../util/interaction';
 import { AnimateComponent } from '../animation/animate-component';
 import { DefaultAxisAnimation } from './animate/config';
+import type { IBaseScale } from '@visactor/vscale';
 
 export abstract class AxisBase<T extends AxisBaseAttributes> extends AnimateComponent<Required<T>> {
   name = 'axis';
+
+  lastScale: IBaseScale;
 
   // TODO: 组件整体统一起来
   protected _innerView: IGroup;
@@ -118,8 +121,8 @@ export abstract class AxisBase<T extends AxisBaseAttributes> extends AnimateComp
    */
   getBoundsWithoutRender(attributes: Partial<T>) {
     const currentAttribute = cloneDeep(this.attribute);
-    // lastScale 不能拷贝
-    currentAttribute.lastScale = (this.attribute as any).lastScale;
+    // scale 不能拷贝
+    currentAttribute.scale = (this.attribute as any).scale;
     merge(this.attribute, attributes);
 
     const offscreenGroup = graphicCreator.group({
@@ -146,10 +149,6 @@ export abstract class AxisBase<T extends AxisBaseAttributes> extends AnimateComp
     this._bindEvent();
     // 尝试执行动画
     this.runAnimation();
-  }
-
-  protected runAnimation() {
-    return;
   }
 
   protected _prepare() {
@@ -568,6 +567,109 @@ export abstract class AxisBase<T extends AxisBaseAttributes> extends AnimateComp
       });
     });
     return data;
+  }
+
+  protected runAnimation() {
+    const lastScale = this.lastScale;
+    if ((this.attribute as any).scale) {
+      const scale = (this.attribute as any).scale;
+      this.lastScale = scale.clone();
+      this.lastScale.range([0, 1]);
+    }
+
+    if (this.attribute.animation && (this as any).applyAnimationState) {
+      // @ts-ignore
+      const currentInnerView = this.getInnerView();
+      // @ts-ignore
+      const prevInnerView = this.getPrevInnerView();
+      if (!prevInnerView) {
+        return;
+      }
+
+      const animationConfig = this._animationConfig;
+
+      this._newElementAttrMap = {};
+
+      // 遍历新的场景树，将新节点属性更新为旧节点
+      // TODO: 目前只处理更新场景
+      traverseGroup(currentInnerView, (el: IGraphic) => {
+        if ((el as IGraphic).type !== 'group' && el.id) {
+          const oldEl = prevInnerView[el.id];
+          // 删除旧图元的动画
+          el.setFinalAttribute(el.attribute);
+          if (oldEl) {
+            oldEl.release();
+            // oldEl.stopAnimationState('enter');
+            // oldEl.stopAnimationState('update');
+            const oldAttrs = (oldEl as IGraphic).attribute;
+            const finalAttrs = el.getFinalAttribute();
+            const diffAttrs: Record<string, any> = diff(oldAttrs, finalAttrs);
+
+            let hasDiff = Object.keys(diffAttrs).length > 0;
+            // TODO 如果入场会有fadeIn，则需要处理opacity（后续还需要考虑其他动画情况）
+            if ('opacity' in oldAttrs && finalAttrs.opacity !== oldAttrs.opacity) {
+              diffAttrs.opacity = finalAttrs.opacity ?? 1;
+              hasDiff = true;
+            }
+
+            if (animationConfig.update && hasDiff) {
+              this._newElementAttrMap[el.id] = {
+                state: 'update',
+                node: el,
+                attrs: el.attribute
+              };
+              const oldAttrs = (oldEl as IGraphic).attribute;
+
+              (el as IGraphic).setAttributes(oldAttrs);
+
+              el.applyAnimationState(
+                ['update'],
+                [
+                  {
+                    name: 'update',
+                    animation: {
+                      selfOnly: true,
+                      ...animationConfig.update,
+                      type: 'axisUpdate',
+                      customParameters: {
+                        config: animationConfig.update,
+                        diffAttrs,
+                        lastScale
+                      }
+                    }
+                  }
+                ]
+              );
+            }
+          } else if (animationConfig.enter) {
+            this._newElementAttrMap[el.id] = {
+              state: 'enter',
+              node: el,
+              attrs: el.attribute
+            };
+
+            el.applyAnimationState(
+              ['enter'],
+              [
+                {
+                  name: 'enter',
+                  animation: {
+                    ...animationConfig.enter,
+                    type: 'axisEnter',
+                    selfOnly: true,
+                    customParameters: {
+                      config: animationConfig.enter,
+                      lastScale,
+                      getTickCoord: this.getTickCoord.bind(this)
+                    }
+                  }
+                }
+              ]
+            );
+          }
+        }
+      });
+    }
   }
 
   release(): void {
