+++ packages/vrender-components/src/axis/line.ts
@@ -46,7 +46,6 @@ export interface LineAxis
 
 export class LineAxis extends AxisBase<LineAxisAttributes> {
   static defaultAttributes = DEFAULT_AXIS_THEME;
-  lastScale: any;
 
   constructor(attributes: LineAxisAttributes, options?: ComponentOptions) {
     super(options?.skipDefault ? attributes : merge({}, LineAxis.defaultAttributes, attributes), options);
@@ -658,114 +657,6 @@ export class LineAxis extends AxisBase<LineAxisAttributes> {
       limitLength = (limitLength - labelSpace - titleSpacing - titleHeight - axisLineWidth - tickLength) / layerCount;
     }
     return limitLength;
-  }
-
-  protected runAnimation() {
-    const lastScale = this.lastScale;
-    if ((this.attribute as any).lastScale) {
-      const lastScale = (this.attribute as any).lastScale;
-      this.lastScale = lastScale.clone();
-      if (lastScale._niceType) {
-        this.lastScale._niceType = lastScale._niceType;
-        this.lastScale._domainValidator = lastScale._domainValidator;
-        this.lastScale._niceDomain = lastScale._niceDomain.slice();
-        this.lastScale.range([0, 1]);
-      }
-    }
-
-    if (this.attribute.animation && (this as any).applyAnimationState) {
-      // @ts-ignore
-      const currentInnerView = this.getInnerView();
-      // @ts-ignore
-      const prevInnerView = this.getPrevInnerView();
-      if (!prevInnerView) {
-        return;
-      }
-
-      const animationConfig = this._animationConfig;
-
-      this._newElementAttrMap = {};
-
-      // 遍历新的场景树，将新节点属性更新为旧节点
-      // TODO: 目前只处理更新场景
-      traverseGroup(currentInnerView, (el: IGraphic) => {
-        if ((el as IGraphic).type !== 'group' && el.id) {
-          const oldEl = prevInnerView[el.id];
-          // 删除旧图元的动画
-          el.setFinalAttribute(el.attribute);
-          if (oldEl) {
-            oldEl.release();
-            // oldEl.stopAnimationState('enter');
-            // oldEl.stopAnimationState('update');
-            const oldAttrs = (oldEl as IGraphic).attribute;
-            const finalAttrs = el.getFinalAttribute();
-            const diffAttrs: Record<string, any> = diff(oldAttrs, finalAttrs);
-
-            let hasDiff = Object.keys(diffAttrs).length > 0;
-            // TODO 如果入场会有fadeIn，则需要处理opacity（后续还需要考虑其他动画情况）
-            if ('opacity' in oldAttrs && finalAttrs.opacity !== oldAttrs.opacity) {
-              diffAttrs.opacity = finalAttrs.opacity ?? 1;
-              hasDiff = true;
-            }
-
-            if (animationConfig.update && hasDiff) {
-              this._newElementAttrMap[el.id] = {
-                state: 'update',
-                node: el,
-                attrs: el.attribute
-              };
-              const oldAttrs = (oldEl as IGraphic).attribute;
-
-              (el as IGraphic).setAttributes(oldAttrs);
-
-              el.applyAnimationState(
-                ['update'],
-                [
-                  {
-                    name: 'update',
-                    animation: {
-                      selfOnly: true,
-                      ...animationConfig.update,
-                      type: 'axisUpdate',
-                      customParameters: {
-                        config: animationConfig.update,
-                        diffAttrs,
-                        lastScale
-                      }
-                    }
-                  }
-                ]
-              );
-            }
-          } else if (animationConfig.enter) {
-            this._newElementAttrMap[el.id] = {
-              state: 'enter',
-              node: el,
-              attrs: el.attribute
-            };
-
-            el.applyAnimationState(
-              ['enter'],
-              [
-                {
-                  name: 'enter',
-                  animation: {
-                    ...animationConfig.enter,
-                    type: 'axisEnter',
-                    selfOnly: true,
-                    customParameters: {
-                      config: animationConfig.enter,
-                      lastScale,
-                      getTickCoord: this.getTickCoord.bind(this)
-                    }
-                  }
-                }
-              ]
-            );
-          }
-        }
-      });
-    }
   }
 
   release(): void {
